Ajustes

+++ LojaDoSeuManoel.API/Controllers/PedidoController.cs
@@ -1,4 +1,5 @@
-﻿using LojaDoSeuManoel.Application.DTOs;
+﻿using LojaDoSeuManoel.API.Security;
+using LojaDoSeuManoel.Application.DTOs;
 using LojaDoSeuManoel.Application.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -20,6 +21,11 @@ namespace LojaDoSeuManoel.API.Controllers
         [ProducesResponseType(typeof(PedidoResponseDTO), 201)]
         public async Task<IActionResult> Post([FromBody] PedidoRequestDTO request)
         {
+            if (!BasicAuthenticationSecurity.AuthenticateRequest(Request))
+            {
+                return Unauthorized(new { Message = "Autenticação falhou. Credenciais inválidas." });
+            }
+
             return StatusCode(201,await _pedidoService.AddAsync(request));
         }
 
@@ -27,6 +33,11 @@ namespace LojaDoSeuManoel.API.Controllers
         [ProducesResponseType(typeof(PedidoResponseDTO), 200)]
         public async Task<IActionResult> Put(Guid id,[FromBody] PedidoRequestDTO request)
         {
+            if (!BasicAuthenticationSecurity.AuthenticateRequest(Request))
+            {
+                return Unauthorized(new { Message = "Autenticação falhou. Credenciais inválidas." });
+            }
+
             return StatusCode(200, await _pedidoService.UpdateAsync(request));
         }
 
@@ -34,6 +45,11 @@ namespace LojaDoSeuManoel.API.Controllers
         [ProducesResponseType(typeof(PedidoResponseDTO), 200)]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!BasicAuthenticationSecurity.AuthenticateRequest(Request))
+            {
+                return Unauthorized(new { Message = "Autenticação falhou. Credenciais inválidas." });
+            }
+
             return StatusCode(200,await _pedidoService.DeleteAsync(id));
         }       
 
@@ -41,6 +57,11 @@ namespace LojaDoSeuManoel.API.Controllers
         [ProducesResponseType(typeof(PedidoResponseDTO), 200)]
         public async Task<IActionResult> GetById(int id)
         {
+            if (!BasicAuthenticationSecurity.AuthenticateRequest(Request))
+            {
+                return Unauthorized(new { Message = "Autenticação falhou. Credenciais inválidas." });
+            }
+
             return StatusCode(200,await _pedidoService.GetByIdAsync(id));
         }
